# 384 speed up dataset indexing pw