--- a/lightly/data/_helpers.py
+++ b/lightly/data/_helpers.py
@@ -33,9 +33,7 @@ def _contains_videos(root: str, extensions: tuple):
         True if root contains subdirectories else false.
     """
     with os.scandir(root) as scan_dir:
-        return any(
-            [f.name.lower().endswith(extensions) for f in scan_dir]
-        )
+        return any(f.name.lower().endswith(extensions) for f in scan_dir)
 
 
 def _is_lightly_output_dir(dirname: str):
@@ -62,9 +60,8 @@ def _contains_subdirs(root: str):
 
     """
     with os.scandir(root) as scan_dir:
-        return any(
-            [f.is_dir() for f in scan_dir if not _is_lightly_output_dir(f.name)]
-        )
+        return any(f.is_dir() for f in scan_dir \
+            if not _is_lightly_output_dir(f.name))
 
 
 def _load_dataset_from_folder(root: str, transform):