# 384 speed up dataset indexing

Closes #384.

I made use of `os.scandir` to get a speed-up of approximately 2.

| Dataset | Images | Before | After |
|----------|---------|---------|------|
| `bdd100k/train` | 70000 | 500ms | 270ms |

The biggest speed-up comes from using `os.scandir` when checking for subdirectories. As described [here](https://docs.python.org/3/library/os.html#os.scandir), `os.scandir` can "significantly increase the performance of code that also needs file type or file attribute information".